+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Updated the tutorial (text and code) https://github.com/Textualize/textual/pull/5257
+- The Select widget no longer scrolls it's container when expanded
 
 ### Fixed